+++ packages/react/FormField/src/components/FormFieldInput/formFieldInput.types.ts
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from "react";
 
-import { InputProps } from "@nimbus-ds/input";
+import { InputProps } from "@nimbus-ds/components";
 import { FormFieldProps } from "../../formField.types";
 
 export type FormFieldInputProps = Omit<FormFieldProps, "children"> &