+++ packages/react/FormField/src/components/FormFieldSelect/formFieldSelect.types.ts
@@ -1,6 +1,6 @@
 import { HTMLAttributes } from "react";
 
-import { SelectProps } from "@nimbus-ds/select";
+import { SelectProps } from "@nimbus-ds/components";
 import { FormFieldProps } from "../../formField.types";
 
 export type FormFieldSelectProps = Omit<FormFieldProps, "children"> &