+++ packages/react/Menu/src/Menu.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Box } from "@nimbus-ds/box";
+import { Box } from "@nimbus-ds/components";
 
 import { MenuButton } from "@nimbus-ds/menubutton";