+++ packages/react/AppShell/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 The AppShell component is the main outer frame of an application. It provides the basic architecture to build an application inside of our admin.
 
+## 2023-02-17 `1.1.0`
+
+### 🎉 New features
+
+- Removed external dependency from `@tiendanube/icons` package to now use internal `@nimbus-ds/icons` package. ([#18](https://github.com/TiendaNube/nimbus-design-system/pull/#18) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 📚 3rd party library updates
+
+- Added `@nimbus-ds/components@2.8.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `@nimbus-ds/styles@6.11.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `@nimbus-ds/icons@1.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/button@2.1.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/icon@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/icon-button@3.3.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/tag@2.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@nimbus-ds/text@5.0.0`. ([#18](https://github.com/TiendaNube/nimbus-patterns/pull/18) by [@juniorconquista](https://github.com/juniorconquista))
+- Removed `@tiendanube/icons@0.3.1`. ([#18](https://github.com/TiendaNube/nimbus-design-system/pull/#18) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-14 `1.0.1`
 
 ### 💡 Others
@@ -21,4 +39,3 @@ The AppShell component is the main outer frame of an application. It provides th
 - Added stories documentation for `AppShell` component. ([#16](https://github.com/TiendaNube/nimbus-patterns/pull/16) by [@juanchigallego](https://github.com/juanchigallego))
 - Added new `AppShell.Header` subcomponent. ([#16](https://github.com/TiendaNube/nimbus-patterns/pull/16) by [@juanchigallego](https://github.com/juanchigallego))
 - Added `leftSlot` and `rightSlot` properties to the `AppShell.Header` subcomponent API. ([#16](https://github.com/TiendaNube/nimbus-patterns/pull/16) by [@juanchigallego](https://github.com/juanchigallego))
-